refactor: decouple tree from routes
