+++ app/entry.server.tsx
@@ -7,9 +7,8 @@ import {
 	ColorModeContext,
 	getColorModeCookie,
 	getInitialColorModeCookie,
-	createEmotionCache,
-	EmotionServerContext,
-} from "~tree";
+} from "~app/colormode";
+import { createEmotionCache, EmotionServerContext } from "~app/emotion";
 
 const cache = createEmotionCache(); // @todo Figure out if global style caching is a good idea
 // eslint-disable-next-line @typescript-eslint/unbound-method
